move get_selected_value to utils
