+++ tests/test_file_data.py
@@ -3729,6 +3729,27 @@ class IndCQCDataUtils:
         ("1-000003", 4.0, "already_populated"),
     ]
 
+    test_first_selection_rows = [
+        ("loc 1", 1, None, 100.0),
+        ("loc 1", 2, 2.0, 50.0),
+        ("loc 1", 3, 3.0, 25.0),
+    ]
+    expected_test_first_selection_rows = [
+        ("loc 1", 1, None, 100.0, 50.0),
+        ("loc 1", 2, 2.0, 50.0, 50.0),
+        ("loc 1", 3, 3.0, 25.0, 50.0),
+    ]
+    test_last_selection_rows = [
+        ("loc 1", 1, 1.0, 100.0),
+        ("loc 1", 2, 2.0, 50.0),
+        ("loc 1", 3, None, 25.0),
+    ]
+    expected_test_last_selection_rows = [
+        ("loc 1", 1, 1.0, 100.0, 50.0),
+        ("loc 1", 2, 2.0, 50.0, 50.0),
+        ("loc 1", 3, None, 25.0, 50.0),
+    ]
+
 
 @dataclass
 class CleanIndCQCData:
@@ -4991,47 +5012,6 @@ class ModelExtrapolationNew:
         ("1-002", 1675209600, 4.0, 1675209600, 1675209600, 4.0, 4.0, None),
         ("1-002", 1677628800, None, 1675209600, 1675209600, 5.0, 6.0, 5.0),
         ("1-003", 1677628800, None, None, None, None, None, None),
-    ]
-
-    test_min_selection_rows = [
-        ("loc 1", 1, 1.0, 100.0),
-        ("loc 1", 2, 2.0, 50.0),
-        ("loc 1", 3, None, 25.0),
-    ]
-    expected_test_min_selection_rows = [
-        ("loc 1", 1, 1.0, 100.0, 50.0),
-        ("loc 1", 2, 2.0, 50.0, 50.0),
-        ("loc 1", 3, None, 25.0, 50.0),
-    ]
-    test_max_selection_rows = [
-        ("loc 1", 1, 1.0, 100.0),
-        ("loc 1", 2, 2.0, 50.0),
-        ("loc 1", 3, None, 25.0),
-    ]
-    expected_test_max_selection_rows = [
-        ("loc 1", 1, 1.0, 100.0, 100.0),
-        ("loc 1", 2, 2.0, 50.0, 100.0),
-        ("loc 1", 3, None, 25.0, 100.0),
-    ]
-    test_first_selection_rows = [
-        ("loc 1", 1, None, 100.0),
-        ("loc 1", 2, 2.0, 50.0),
-        ("loc 1", 3, 3.0, 25.0),
-    ]
-    expected_test_first_selection_rows = [
-        ("loc 1", 1, None, 100.0, 50.0),
-        ("loc 1", 2, 2.0, 50.0, 50.0),
-        ("loc 1", 3, 3.0, 25.0, 50.0),
-    ]
-    test_last_selection_rows = [
-        ("loc 1", 1, 1.0, 100.0),
-        ("loc 1", 2, 2.0, 50.0),
-        ("loc 1", 3, None, 25.0),
-    ]
-    expected_test_last_selection_rows = [
-        ("loc 1", 1, 1.0, 100.0, 50.0),
-        ("loc 1", 2, 2.0, 50.0, 50.0),
-        ("loc 1", 3, None, 25.0, 50.0),
     ]
