+++ tests/unit/test_extrapolation_new.py
@@ -322,77 +322,3 @@ class CombineExtrapolationTests(ModelExtrapolationTests):
                 self.expected_data[i][self.extrapolation_model_name],
                 f"Returned value in row {i} does not match expected",
             )
-
-
-class GetSelectedValueFunctionTests(ModelExtrapolationTests):
-    def setUp(self):
-        super().setUp()
-        self.w = (
-            Window.partitionBy(IndCqc.location_id)
-            .orderBy(IndCqc.unix_time)
-            .rowsBetween(Window.unboundedPreceding, Window.unboundedFollowing)
-        )
-
-    def test_get_selected_value_returns_correct_values_when_selection_equals_first(
-        self,
-    ):
-        test_df = self.spark.createDataFrame(
-            Data.test_first_selection_rows, Schemas.get_selected_value_schema
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_test_first_selection_rows,
-            Schemas.expected_get_selected_value_schema,
-        )
-        returned_df = job.get_selected_value(
-            test_df,
-            self.w,
-            IndCqc.ascwds_filled_posts_dedup_clean,
-            IndCqc.rolling_average_model,
-            "new_column",
-            selection="first",
-        )
-        self.assertEqual(
-            returned_df.sort(IndCqc.location_id, IndCqc.unix_time).collect(),
-            expected_df.collect(),
-        )
-
-    def test_get_selected_value_returns_correct_values_when_selection_equals_last(self):
-        test_df = self.spark.createDataFrame(
-            Data.test_last_selection_rows, Schemas.get_selected_value_schema
-        )
-        expected_df = self.spark.createDataFrame(
-            Data.expected_test_last_selection_rows,
-            Schemas.expected_get_selected_value_schema,
-        )
-        returned_df = job.get_selected_value(
-            test_df,
-            self.w,
-            IndCqc.ascwds_filled_posts_dedup_clean,
-            IndCqc.rolling_average_model,
-            "new_column",
-            selection="last",
-        )
-        self.assertEqual(
-            returned_df.sort(IndCqc.location_id, IndCqc.unix_time).collect(),
-            expected_df.collect(),
-        )
-
-    def test_get_selected_value_raises_error_when_selection_is_not_permitted(self):
-        test_df = self.spark.createDataFrame(
-            Data.test_last_selection_rows, Schemas.get_selected_value_schema
-        )
-
-        with self.assertRaises(ValueError) as context:
-            job.get_selected_value(
-                test_df,
-                self.w,
-                IndCqc.ascwds_filled_posts_dedup_clean,
-                IndCqc.rolling_average_model,
-                "new_column",
-                selection="other",
-            )
-
-        self.assertTrue(
-            "Error: The selection parameter 'other' was not found. Please use 'first' or 'last'.",
-            "Exception does not contain the correct error message",
-        )
