+++ utils/ind_cqc_filled_posts_utils/utils.py
@@ -1,4 +1,4 @@
-from pyspark.sql import DataFrame, functions as F
+from pyspark.sql import DataFrame, functions as F, Window
 from pyspark.sql.types import (
     IntegerType,
     StringType,
@@ -54,5 +54,53 @@ def populate_estimate_filled_posts_and_source_in_the_order_of_the_column_list(
             IndCQC.estimate_filled_posts_source,
             model_name,
         )
+
+    return df
+
+
+def get_selected_value(
+    df: DataFrame,
+    window_spec: Window,
+    column_with_null_values: str,
+    column_with_data: str,
+    new_column: str,
+    selection: str,
+) -> DataFrame:
+    """
+    Creates a new column with the selected value (first or last) from a given column.
+
+    This function creates a new column by selecting a specified value over a given window on a given dataframe. It will
+    only select values in the column with data that have null values in the original column.
+
+    Args:
+        df (DataFrame): A dataframe containing the supplied columns.
+        window_spec (Window): A window describing how to prepare the dataframe.
+        column_with_null_values (str): A column with missing data.
+        column_with_data (str): A column with data for all the rows that column_with_null_values has data. This can be column_with_null_values itself.
+        new_column (str): The name of the new column containing the resulting selected values.
+        selection (str): One of 'first' or 'last'. This determines which pyspark window function will be used.
+
+    Returns:
+        DataFrame: A dataframe containing a new column with the selected value populated through each window.
+
+    Raises:
+        ValueError: If 'selection' is not one of the two permitted pyspark window functions.
+    """
+    selection_methods = {"first": F.first, "last": F.last}
+
+    if selection not in selection_methods:
+        raise ValueError(
+            f"Error: The selection parameter '{selection}' was not found. Please use 'first' or 'last'."
+        )
+
+    method = selection_methods[selection]
+
+    df = df.withColumn(
+        new_column,
+        method(
+            F.when(F.col(column_with_null_values).isNotNull(), F.col(column_with_data)),
+            ignorenulls=True,
+        ).over(window_spec),
+    )
 
     return df
